refactor(ui): split task form leaf components
